Rescale annual heat demand to resolution
